# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface Note {
  id: string;
  userId: string;
  title: string;
  content: string;
  color: string;
  pinned: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface NoteStats {
  totalNotes: number;
  pinnedNotes: number;
  updatedToday: number;
}

export interface CreateNoteBody {
  title?: string;
  content?: string;
  color?: string;
  pinned?: boolean;
}

export interface UpdateNoteBody {
  title?: string;
  content?: string;
  color?: string;
  pinned?: boolean;
}

export interface HealthStatus {
  status: string;
}

export interface Recording {
  id: string;
  userId: string;
  title: string;
  description?: string | null;
  /** Duration in seconds */
  duration: number;
  /** Path to audio file in object storage */
  audioObjectPath: string;
  /** Unique token for public sharing */
  shareToken: string;
  createdAt: string;
  updatedAt: string;
}

export interface RecordingStats {
  totalTapes: number;
  totalDurationSeconds: number;
  newestTapeDate?: string | null;
}

export interface CreateRecordingBody {
  title: string;
  description?: string | null;
  duration: number;
  audioObjectPath: string;
}

export interface UpdateRecordingBody {
  title?: string;
  description?: string | null;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListNotesParams = {
/**
 * Search query to filter notes by title or content
 */
q?: string;
};