# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: recordings
    description: Voice memo recordings
  - name: storage
    description: Object storage operations
  - name: notes
    description: Note-taking operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /recordings:
    get:
      operationId: listRecordings
      tags: [recordings]
      summary: List all recordings for the authenticated user
      responses:
        "200":
          description: List of recordings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recording"
        "401":
          description: Unauthorized
    post:
      operationId: createRecording
      tags: [recordings]
      summary: Create a new recording
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecordingBody"
      responses:
        "201":
          description: Created recording
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
        "401":
          description: Unauthorized

  /recordings/stats:
    get:
      operationId: getRecordingStats
      tags: [recordings]
      summary: Get recording statistics for the authenticated user
      responses:
        "200":
          description: Recording stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingStats"
        "401":
          description: Unauthorized

  /recordings/{id}:
    get:
      operationId: getRecording
      tags: [recordings]
      summary: Get a single recording (must be owner)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recording
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    patch:
      operationId: updateRecording
      tags: [recordings]
      summary: Update recording label (title, description)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecordingBody"
      responses:
        "200":
          description: Updated recording
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteRecording
      tags: [recordings]
      summary: Delete a recording
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /shared/{shareToken}:
    get:
      operationId: getSharedRecording
      tags: [recordings]
      summary: Get a publicly shared recording by share token (no auth required)
      parameters:
        - name: shareToken
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shared recording
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
        "404":
          description: Not found

  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all notes for the authenticated user
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search query to filter notes by title or content
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
        "401":
          description: Unauthorized
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "401":
          description: Unauthorized

  /notes/stats:
    get:
      operationId: getNoteStats
      tags: [notes]
      summary: Get note statistics for the authenticated user
      responses:
        "200":
          description: Note stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteStats"
        "401":
          description: Unauthorized

  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a single note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteBody"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned URL for uploading audio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
        "401":
          description: Unauthorized

  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [storage]
      summary: Serve a stored object (audio file)
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
        "404":
          description: Not found

components:
  schemas:
    Note:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        content:
          type: string
        color:
          type: string
        pinned:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - content
        - color
        - pinned
        - createdAt
        - updatedAt

    NoteStats:
      type: object
      properties:
        totalNotes:
          type: number
        pinnedNotes:
          type: number
        updatedToday:
          type: number
      required:
        - totalNotes
        - pinnedNotes
        - updatedToday

    CreateNoteBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        color:
          type: string
        pinned:
          type: boolean

    UpdateNoteBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        color:
          type: string
        pinned:
          type: boolean

    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Recording:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        duration:
          type: number
          description: Duration in seconds
        audioObjectPath:
          type: string
          description: Path to audio file in object storage
        shareToken:
          type: string
          description: Unique token for public sharing
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - duration
        - audioObjectPath
        - shareToken
        - createdAt
        - updatedAt

    RecordingStats:
      type: object
      properties:
        totalTapes:
          type: number
        totalDurationSeconds:
          type: number
        newestTapeDate:
          type: string
          format: date-time
          nullable: true
      required:
        - totalTapes
        - totalDurationSeconds

    CreateRecordingBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        duration:
          type: number
        audioObjectPath:
          type: string
      required:
        - title
        - duration
        - audioObjectPath

    UpdateRecordingBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true

    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: number
        contentType:
          type: string
      required:
        - name
        - size
        - contentType

    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required:
        - uploadURL
        - objectPath