# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary List all recordings for the authenticated user
 */
export const ListRecordingsResponseItem = zod.object({
  "id": zod.string(),
  "userId": zod.string(),
  "title": zod.string(),
  "description": zod.string().nullish(),
  "duration": zod.number().describe('Duration in seconds'),
  "audioObjectPath": zod.string().describe('Path to audio file in object storage'),
  "shareToken": zod.string().describe('Unique token for public sharing'),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date()
})
export const ListRecordingsResponse = zod.array(ListRecordingsResponseItem)


/**
 * @summary Create a new recording
 */
export const CreateRecordingBody = zod.object({
  "title": zod.string(),
  "description": zod.string().nullish(),
  "duration": zod.number(),
  "audioObjectPath": zod.string()
})


/**
 * @summary Get recording statistics for the authenticated user
 */
export const GetRecordingStatsResponse = zod.object({
  "totalTapes": zod.number(),
  "totalDurationSeconds": zod.number(),
  "newestTapeDate": zod.coerce.date().nullish()
})


/**
 * @summary Get a single recording (must be owner)
 */
export const GetRecordingParams = zod.object({
  "id": zod.coerce.string()
})

export const GetRecordingResponse = zod.object({
  "id": zod.string(),
  "userId": zod.string(),
  "title": zod.string(),
  "description": zod.string().nullish(),
  "duration": zod.number().describe('Duration in seconds'),
  "audioObjectPath": zod.string().describe('Path to audio file in object storage'),
  "shareToken": zod.string().describe('Unique token for public sharing'),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date()
})


/**
 * @summary Update recording label (title, description)
 */
export const UpdateRecordingParams = zod.object({
  "id": zod.coerce.string()
})

export const UpdateRecordingBody = zod.object({
  "title": zod.string().optional(),
  "description": zod.string().nullish()
})

export const UpdateRecordingResponse = zod.object({
  "id": zod.string(),
  "userId": zod.string(),
  "title": zod.string(),
  "description": zod.string().nullish(),
  "duration": zod.number().describe('Duration in seconds'),
  "audioObjectPath": zod.string().describe('Path to audio file in object storage'),
  "shareToken": zod.string().describe('Unique token for public sharing'),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date()
})


/**
 * @summary Delete a recording
 */
export const DeleteRecordingParams = zod.object({
  "id": zod.coerce.string()
})


/**
 * @summary Get a publicly shared recording by share token (no auth required)
 */
export const GetSharedRecordingParams = zod.object({
  "shareToken": zod.coerce.string()
})

export const GetSharedRecordingResponse = zod.object({
  "id": zod.string(),
  "userId": zod.string(),
  "title": zod.string(),
  "description": zod.string().nullish(),
  "duration": zod.number().describe('Duration in seconds'),
  "audioObjectPath": zod.string().describe('Path to audio file in object storage'),
  "shareToken": zod.string().describe('Unique token for public sharing'),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date()
})


/**
 * @summary List all notes for the authenticated user
 */
export const ListNotesQueryParams = zod.object({
  "q": zod.coerce.string().optional().describe('Search query to filter notes by title or content')
})

export const ListNotesResponseItem = zod.object({
  "id": zod.string(),
  "userId": zod.string(),
  "title": zod.string(),
  "content": zod.string(),
  "color": zod.string(),
  "pinned": zod.boolean(),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date()
})
export const ListNotesResponse = zod.array(ListNotesResponseItem)


/**
 * @summary Create a new note
 */
export const CreateNoteBody = zod.object({
  "title": zod.string().optional(),
  "content": zod.string().optional(),
  "color": zod.string().optional(),
  "pinned": zod.boolean().optional()
})


/**
 * @summary Get note statistics for the authenticated user
 */
export const GetNoteStatsResponse = zod.object({
  "totalNotes": zod.number(),
  "pinnedNotes": zod.number(),
  "updatedToday": zod.number()
})


/**
 * @summary Get a single note
 */
export const GetNoteParams = zod.object({
  "id": zod.coerce.string()
})

export const GetNoteResponse = zod.object({
  "id": zod.string(),
  "userId": zod.string(),
  "title": zod.string(),
  "content": zod.string(),
  "color": zod.string(),
  "pinned": zod.boolean(),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date()
})


/**
 * @summary Update a note
 */
export const UpdateNoteParams = zod.object({
  "id": zod.coerce.string()
})

export const UpdateNoteBody = zod.object({
  "title": zod.string().optional(),
  "content": zod.string().optional(),
  "color": zod.string().optional(),
  "pinned": zod.boolean().optional()
})

export const UpdateNoteResponse = zod.object({
  "id": zod.string(),
  "userId": zod.string(),
  "title": zod.string(),
  "content": zod.string(),
  "color": zod.string(),
  "pinned": zod.boolean(),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date()
})


/**
 * @summary Delete a note
 */
export const DeleteNoteParams = zod.object({
  "id": zod.coerce.string()
})


/**
 * @summary Request a presigned URL for uploading audio
 */
export const RequestUploadUrlBody = zod.object({
  "name": zod.string(),
  "size": zod.number(),
  "contentType": zod.string()
})

export const RequestUploadUrlResponse = zod.object({
  "uploadURL": zod.string(),
  "objectPath": zod.string()
})


/**
 * @summary Serve a stored object (audio file)
 */
export const GetStorageObjectParams = zod.object({
  "objectPath": zod.coerce.string()
})